Draw random rectangles

+++ src/elos/kernel/common/types.h
@@ -2,12 +2,12 @@
 
 // We assume 64-bit system
 
-typedef unsigned long int u64;
-typedef unsigned int      u32;
-typedef unsigned short    u16;
-typedef unsigned char     u8;
+typedef unsigned long long u64;
+typedef unsigned int       u32;
+typedef unsigned short     u16;
+typedef unsigned char      u8;
 
-typedef long int          s64;
+typedef long long         s64;
 typedef int               s32;
 typedef short             s16;
 typedef char              s8;
